moved ListenSubmissionState to shared module under androidMain and updated its usage in ListenServiceManager with shared notification manager and shared work manager
